+++ pallets/pallet-did-lookup/src/tests/delete.rs
