+++ pallets/pallet-web3-names/src/tests/mod.rs
@@ -15,6 +15,8 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
-mod create;
-mod delete;
+
+mod ban;
+mod claim;
 mod deposit;
+mod release;
